+++ UtilityBelt/Lib/Constants/Skills.cs
@@ -5,41 +5,60 @@ using System.Text;
 
 namespace UtilityBelt.Lib.Constants {
      public enum Skills : uint {
-        Axe = 0x01,
-        Bow = 0x02,
-        Crossbow = 0x03,
-        Dagger = 0x04,
-        Mace = 0x05,
-        Melee_Defense = 0x06,
-        Missile_Defense = 0x07,
-        Spear = 0x09,
-        Staff = 0x0A,
-        Sword = 0x0B,
-        Thrown_Weapons = 0x0C,
-        Unarmed_Combat = 0x0D,
-        Arcane_Lore = 0x0E,
-        Magic_Defense = 0x0F,
-        Mana_Conversion = 0x10,
-        Item_Tinkering = 0x12,
-        Assess_Person = 0x13,
-        Deception = 0x14,
-        Healing = 0x15,
-        Jump = 0x16,
-        Lockpick = 0x17,
-        Run = 0x18,
-        Assess_Creature = 0x1B,
-        Weapon_Tinkering = 0x1C,
-        Armor_Tinkering = 0x1D,
-        Magic_Item_Tinkering = 0x1E,
-        Creature_Enchantment = 0x1F,
-        Item_Enchantment = 0x20,
-        Life_Magic = 0x21,
-        War_Magic = 0x22,
-        Leadership = 0x23,
-        Loyalty = 0x24,
-        Fletching = 0x25,
-        Alchemy = 0x26,
-        Cooking = 0x27,
-        Salvaging = 0x28
+        None,
+        Axe,                 /* Retired */
+        Bow,                 /* Retired */
+        Crossbow,            /* Retired */
+        Dagger,              /* Retired */
+        Mace,                /* Retired */
+        MeleeDefense,
+        MissileDefense,
+        Sling,               /* Retired */
+        Spear,               /* Retired */
+        Staff,               /* Retired */
+        Sword,               /* Retired */
+        ThrownWeapon,        /* Retired */
+        UnarmedCombat,       /* Retired */
+        ArcaneLore,
+        MagicDefense,
+        ManaConversion,
+        Spellcraft,          /* Unimplemented */
+        ItemTinkering,
+        AssessPerson,
+        Deception,
+        Healing,
+        Jump,
+        Lockpick,
+        Run,
+        Awareness,           /* Unimplemented */
+        ArmsAndArmorRepair,  /* Unimplemented */
+        AssessCreature,
+        WeaponTinkering,
+        ArmorTinkering,
+        MagicItemTinkering,
+        CreatureEnchantment,
+        ItemEnchantment,
+        LifeMagic,
+        WarMagic,
+        Leadership,
+        Loyalty,
+        Fletching,
+        Alchemy,
+        Cooking,
+        Salvaging,
+        TwoHandedCombat,
+        Gearcraft,           /* Retired */
+        VoidMagic,
+        HeavyWeapons,
+        LightWeapons,
+        FinesseWeapons,
+        MissileWeapons,
+        Shield,
+        DualWield,
+        Recklessness,
+        SneakAttack,
+        DirtyFighting,
+        Challenge,          /* Unimplemented */
+        Summoning
     }
 }